Refactor GardenConfig subclasses into separate classes for readability Accordions/Subcategories
