+++ src/main/java/at/hannibal2/skyhanni/config/features/garden/DicerCounterConfig.java
@@ -1,0 +1,24 @@
+package at.hannibal2.skyhanni.config.features.garden;
+
+import at.hannibal2.skyhanni.config.FeatureToggle;
+import at.hannibal2.skyhanni.config.core.config.Position;
+import com.google.gson.annotations.Expose;
+import io.github.moulberry.moulconfig.annotations.ConfigEditorBoolean;
+import io.github.moulberry.moulconfig.annotations.ConfigOption;
+
+public class DicerCounterConfig {
+    @Expose
+    @ConfigOption(name = "RNG Drop Counter", desc = "Count RNG drops for Melon Dicer and Pumpkin Dicer.")
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean display = true;
+
+    @Expose
+    @ConfigOption(name = "Hide Chat", desc = "Hide the chat message when dropping a RNG Dicer drop.")
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean hideChat = false;
+
+    @Expose
+    public Position pos = new Position(16, -232, false, true);
+}
