+++ src/main/java/at/hannibal2/skyhanni/config/features/garden/FarmingArmorDropsConfig.java
@@ -1,0 +1,24 @@
+package at.hannibal2.skyhanni.config.features.garden;
+
+import at.hannibal2.skyhanni.config.FeatureToggle;
+import at.hannibal2.skyhanni.config.core.config.Position;
+import com.google.gson.annotations.Expose;
+import io.github.moulberry.moulconfig.annotations.ConfigEditorBoolean;
+import io.github.moulberry.moulconfig.annotations.ConfigOption;
+
+public class FarmingArmorDropsConfig {
+    @Expose
+    @ConfigOption(name = "Show Counter", desc = "Count all §9Cropie§7, §5Squash §7and §6Fermento §7dropped.")
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean enabled = true;
+
+    @Expose
+    @ConfigOption(name = "Hide Chat", desc = "Hide the chat message when receiving a farming armor drop.")
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean hideChat = false;
+
+    @Expose
+    public Position pos = new Position(16, -232, false, true);
+}
