+++ src/main/java/at/hannibal2/skyhanni/config/features/garden/cropmilestones/MushroomPetPerkConfig.java
@@ -1,0 +1,43 @@
+package at.hannibal2.skyhanni.config.features.garden.cropmilestones;
+
+import at.hannibal2.skyhanni.config.FeatureToggle;
+import at.hannibal2.skyhanni.config.core.config.Position;
+import com.google.gson.annotations.Expose;
+import io.github.moulberry.moulconfig.annotations.ConfigEditorBoolean;
+import io.github.moulberry.moulconfig.annotations.ConfigEditorDraggableList;
+import io.github.moulberry.moulconfig.annotations.ConfigOption;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+// TODO moulconfig runnable support
+public class MushroomPetPerkConfig {
+    @Expose
+    @ConfigOption(
+        name = "Display Enabled",
+        desc = "Show the progress and ETA for mushroom crops when farming other crops because of the Mooshroom Cow perk.")
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean enabled = true;
+
+    @Expose
+    @ConfigOption(
+        name = "Mushroom Text",
+        desc = "Drag text to change the appearance of the overlay.\n" +
+            "Hold a farming tool to show the overlay."
+    )
+    @ConfigEditorDraggableList(
+        exampleText = {
+            "§6Mooshroom Cow Perk",
+            "§7Mushroom Tier 8",
+            "§e6,700§8/§e15,000",
+            "§7In §b12m 34s",
+            "§7Percentage: §e12.34%",
+        }
+    )
+    public List<Integer> text = new ArrayList<>(Arrays.asList(0, 1, 2, 3));
+
+    @Expose
+    public Position pos = new Position(-112, -143, false, true);
+}
